cleanup and set version to 1003

+++ src/drivers/modem_nf/zephyr/modem_nf.h
@@ -6,6 +6,6 @@ int modem_nf_reset(void);
 int get_pdp_addr(char **);
 int modem_nf_wakeup(void);
 int modem_nf_sleep(void);
-int set_socket_linger_time(uint8_t , int);
+int set_socket_linger_time(int , int);
 int get_ccid(char **);
 #endif /* MODEM_NF_H_ */
